+++ SurfaceTopography/ChangeLog.md
@@ -14,10 +14,12 @@ v0.97.0 (not yet released)
 - ENH: Progress callbacks for some analysis functions
 - ENH: Gaussian process regression for resampling noisy data onto an
   arbitrary grid
-- BUG: Multiple bug fixed in scale-dependent analyses
+- BUG: Multiple bugs fixed in scale-dependent analyses
 - API: Reconciled radial averages into its own regression module;
   default behavior has changed
 - API: All analysis functions now resample data onto a reasonable grid
+- API: Resampling methods no longer discard bins with no data; those are
+  filled with NaNs
 
 v0.96.0 (13Sep21)
 -----------------